# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dance Analysis API",
    "description": "Video dance analysis with pose estimation via microservices",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/sessions/{session_id}/participants": {
      "get": {
        "tags": [
          "session-participants"
        ],
        "summary": "List Participants",
        "description": "List all participants in a session with available users.\n\nRequires: user has access to the session.",
        "operationId": "list_participants_api_v1_sessions__session_id__participants_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "session-participants"
        ],
        "summary": "Create Participant",
        "description": "Create a participant in a session.\n\nThis operation is ATOMIC:\n- Creates participant record\n- Binds to slot (if provided)\n- Enforces one participant per user per session\n- Enforces one user per slot per session\n- Validates role is dancer or coach (v1 only)\n\nRequires: user is the session owner.",
        "operationId": "create_participant_api_v1_sessions__session_id__participants_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionParticipantCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionParticipantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/participants/with-users": {
      "get": {
        "tags": [
          "session-participants"
        ],
        "summary": "List Participants With Users",
        "description": "List participants with their user details.\n\nRequires: user has access to the session.",
        "operationId": "list_participants_with_users_api_v1_sessions__session_id__participants_with_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionParticipantWithUserResponse"
                  },
                  "title": "Response List Participants With Users Api V1 Sessions  Session Id  Participants With Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/available-users": {
      "get": {
        "tags": [
          "session-participants"
        ],
        "summary": "List Available Users",
        "description": "List users available to add as participants.\n\nAvailable = has access + not already a participant + not deleted.\n\nRequires: user has access to the session.",
        "operationId": "list_available_users_api_v1_sessions__session_id__available_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Available Users Api V1 Sessions  Session Id  Available Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/participants/{participant_id}": {
      "get": {
        "tags": [
          "session-participants"
        ],
        "summary": "Get Participant",
        "description": "Get a specific participant.\n\nRequires: user has access to the session.",
        "operationId": "get_participant_api_v1_sessions__session_id__participants__participant_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Participant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionParticipantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "session-participants"
        ],
        "summary": "Update Participant",
        "description": "Update a participant's role and/or slot binding.\n\nCan update:\n- role: dancer or coach\n- dancer_slot_id: bind to slot or null to unbind\n\nEnforces one user per slot constraint.\n\nRequires: user is the session owner.",
        "operationId": "update_participant_api_v1_sessions__session_id__participants__participant_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Participant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionParticipantUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionParticipantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "session-participants"
        ],
        "summary": "Delete Participant",
        "description": "Delete a participant from a session.\n\nRemoves:\n- Participant record\n- Slot binding\n- All associated data\n\nRequires: user is the session owner.",
        "operationId": "delete_participant_api_v1_sessions__session_id__participants__participant_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Participant Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/participants/{participant_id}/bind-slot/{slot_id}": {
      "post": {
        "tags": [
          "session-participants"
        ],
        "summary": "Bind Participant To Slot",
        "description": "Bind a participant to a dancer slot.\n\nEnforces one user per slot per session.\n\nRequires: user is the session owner.",
        "operationId": "bind_participant_to_slot_api_v1_sessions__session_id__participants__participant_id__bind_slot__slot_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Participant Id"
            }
          },
          {
            "name": "slot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Slot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionParticipantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/participants/{participant_id}/unbind-slot": {
      "post": {
        "tags": [
          "session-participants"
        ],
        "summary": "Unbind Participant From Slot",
        "description": "Remove a participant's slot binding.\n\nParticipant remains in session, just no longer bound to a slot.\n\nRequires: user is the session owner.",
        "operationId": "unbind_participant_from_slot_api_v1_sessions__session_id__participants__participant_id__unbind_slot_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Participant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionParticipantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParticipantListResponse": {
        "properties": {
          "participants": {
            "items": {
              "$ref": "#/components/schemas/SessionParticipantResponse"
            },
            "type": "array",
            "title": "Participants"
          },
          "available_users": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Available Users",
            "description": "User IDs with session access but no participant record"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "participants",
          "total_count"
        ],
        "title": "ParticipantListResponse",
        "description": "Response for list of participants with slot information."
      },
...
      "SessionParticipantCreate": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Participant role: dancer or coach",
            "default": "dancer"
          },
          "dancer_slot_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dancer Slot Id",
            "description": "Optional dancer slot to bind participant to"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "SessionParticipantCreate",
        "description": "Request to create a session participant."
      },
      "SessionParticipantResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "dancer_slot_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dancer Slot Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "user_id",
          "role",
          "created_at"
        ],
        "title": "SessionParticipantResponse",
        "description": "Response for session participant record."
      },
      "SessionParticipantUpdate": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role",
            "description": "Participant role: dancer or coach"
          },
          "dancer_slot_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dancer Slot Id",
            "description": "Dancer slot to bind participant to, or null to unbind"
          }
        },
        "type": "object",
        "title": "SessionParticipantUpdate",
        "description": "Request to update a session participant."
      },
      "SessionParticipantWithUserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "dancer_slot_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dancer Slot Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "user_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Email"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "user_id",
          "role",
          "created_at"
        ],
        "title": "SessionParticipantWithUserResponse",
        "description": "Response including participant user details."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}